run-kakao 파일의 모듈화 및 구글 지도에서 바로 복사해서 붙여넣을 수 있도록 수정
